requirements.txt
flask
pillow
opencv-python
